Address comments 2
